screenTime/requirement.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
kiwisolver==1.4.4
matplotlib==3.7.1
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
seaborn==0.12.2
six==1.16.0
tzdata==2023.3
